Update description.ext
Added in an SS Infantry Squad. Somebody please help me unravel this friendly fire problem, it's 1.40 AM here and the SS_InfSquad is set to independent and the others are set to east. When everything was set to east they shoot eachother, when they aren't on east they shoot eachother. I don't know where the problem lies :P